classifier/requirements.txt
kafka-python
influxdb
joblib
sklearn
numpy